+++ reference-impl/index.js
@@ -26,11 +26,12 @@
 'use strict';
 
 const { StegoEncoder, StegoDecoder, MSG_TYPE, MAGIC_BYTE, HEADER_SIZE,
-        VERSION_V1, VERSION_V2, VERSION_V3, isValidHex, safeHexToBuffer
+        VERSION_V1, VERSION_V2, VERSION_V3, isValidHex, safeHexToBuffer,
+        wrapTypedPayload, unwrapTypedPayload
       } = require('./stego-core');
 
 const { E2ECrypto, encryptOneShot, decryptOneShot,
-        ALGORITHM, KEY_LENGTH, IV_LENGTH, AUTH_TAG_LENGTH, SALT_LENGTH, NONCE_LENGTH
+        KEY_LENGTH, AUTH_TAG_LENGTH, SALT_LENGTH, NONCE_LENGTH
       } = require('./crypto');
 
 const { MiningGate, MinerState, ACCESS_LEVEL, DEFAULT_CONFIG } = require('./mining-gate');
@@ -39,10 +40,11 @@ module.exports = {
   // Steganography
   StegoEncoder, StegoDecoder, MSG_TYPE, MAGIC_BYTE, HEADER_SIZE,
   VERSION_V1, VERSION_V2, VERSION_V3, isValidHex, safeHexToBuffer,
+  wrapTypedPayload, unwrapTypedPayload,
 
   // Cryptography
   E2ECrypto, encryptOneShot, decryptOneShot,
-  ALGORITHM, KEY_LENGTH, IV_LENGTH, AUTH_TAG_LENGTH, SALT_LENGTH, NONCE_LENGTH,
+  KEY_LENGTH, AUTH_TAG_LENGTH, SALT_LENGTH, NONCE_LENGTH,
 
   // Mining Gate
   MiningGate, MinerState, ACCESS_LEVEL, DEFAULT_CONFIG